tcl, tk: drop unnecessary patches
Manpage filenames no longer contain colons.

The check for NULL pointer that 006-prevent-tclStubsPtr-segfault does
was added to upstream long time ago. The patch keep applying by
chance.

001-fix-relocation was a deviation from upstream's standard behaviour
to cater to the specific needs of the patch's author for a
non-standard directory layout. It shouldn't never be accepted.

tk/008-dont-link-shared-with--static-libgcc was removed and now
tk/PKGBUILD uses the same find/sed mechanism as in tcl/PKGBUILD for
removing -static-libgcc.